+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 - GCGI-414: Support for multiple requisitions from one donor
 - GCGI-430: Update Mutect2 version to GATK 4.2.6.1
+- GCGI-442: Remove delly results from JSON and HTML
 - Move Djerba version text into assay description section
 
 ## v0.3.3: 2022-07-25